Start developing distinctions between model, view and controller.
Factored as much state as possible out of nested components, drawing it to the App and organizing it with defaults. Similar work on controller logic. Conversion to a more distinct m-v-c model isn't complete, and may not ultimately follow the pattern established with this commit, as I'm still exploring various ways to organize a react project.