Hydrometer correction temperatures adapt to preference changes.
Previously, changes in units of temperature change the units of temperatures entered in the hydrometer correction wizard, but did not change the magnitude of the value. Now, the magnitude of the value changes as well, such that the temperature displayed remains equivalent (within rounding errors) as units of measurement change. Put another way, before, 15C <=> 15F, now 15C <=> 59F.